+++ gemclus/linear/_linear_geminis.py
@@ -348,7 +348,7 @@ class LinearWasserstein(LinearModel):
 
 
 class RIM(LinearModel):
-    """ Implementation of the maximisation of the classical mutual information using a logistic regression with an
+    r""" Implementation of the maximisation of the classical mutual information using a logistic regression with an
         :math:`\ell_2` penalty on the weights. This implementation follows the framework described by Krause et al. in the
         RIM paper.
 
@@ -446,8 +446,8 @@ class RIM(LinearModel):
         self.optimiser_.update_params(weights, gradients)
 
 
-class KernelRIM(DiscriminativeModel):
-    """ Implementation of the maximisation of the classical mutual information using a kernelised version of the
+class KernelRIM(LinearModel):
+    r""" Implementation of the maximisation of the classical mutual information using a kernelised version of the
         logistic regression with an :math:`\ell_2` penalty on the weights. This implementation follows the framework
         described by Krause et al. in the RIM paper.
 
@@ -516,15 +516,13 @@ class KernelRIM(DiscriminativeModel):
         Examples
         --------
         >>> from sklearn.datasets import load_iris
-        >>> from gemclus.linear import RIM
+        >>> from gemclus.linear import KernelRIM
         >>> X,y=load_iris(return_X_y=True)
-        >>> clf = RIM(random_state=0).fit(X)
+        >>> clf = KernelRIM(random_state=0).fit(X)
         >>> clf.predict(X[:2,:])
-        array([0, 0])
+        array([2, 2])
         >>> clf.predict_proba(X[:2,:]).shape
         (2, 3)
-        >>> clf.score(X)
-        0.4390485754
         """
     _parameter_constraints: dict = {
         **DiscriminativeModel._parameter_constraints,
@@ -555,45 +553,31 @@ class KernelRIM(DiscriminativeModel):
         if callable(self.base_kernel):
             if self.base_kernel_params is not None:
                 warnings.warn("Parameters passed through kernel_params are ignored when kernel is a callable.")
-            kernel = self.base_kernel(X)
+            kernel = self.base_kernel(X, self.input_data_)
         else:
             _params = dict() if self.base_kernel_params is None else self.base_kernel_params
             kernel = pairwise_kernels(X, self.input_data_, metric=self.base_kernel, **_params)
         return kernel
-
-    def _init_params(self, random_state, X=None):
-        in_threshold = np.sqrt(1 / self.n_features_in_)
-        self.W_ = random_state.uniform(-in_threshold, in_threshold, size=(len(X), self.n_clusters))
-        self.b_ = random_state.uniform(-in_threshold, in_threshold, size=(1, self.n_clusters))
-
-    def _get_weights(self):
-        return [self.W_, self.b_]
-
-    def _infer(self, X, retain=True):
-        kernel = self._compute_kernel(X)
-        H = kernel @ self.W_ + self.b_
-        return softmax(H)
 
     def fit(self, X, y=None):
         # We start by storing the input data for later kernel computations
         check_array(X)
         self.input_data_ = X
 
-        return super().fit(X, y)
+        training_kernel = self._compute_kernel(X)
+        super().fit(training_kernel, y)
+
+        self.n_features_in_ = X.shape[1]
+
+        return self
 
     def _compute_grads(self, X, y_pred, gradient):
-        tau_hat_grad = y_pred * (gradient - (y_pred * gradient).sum(1, keepdims=True))  # Shape NxK
+        base_grads = super()._compute_grads(X, y_pred, gradient)
+        # Add the regularisation gradient on the weight matrix
+        base_grads[0] += 2 * self.reg * np.dot(X, self.W_)
+        return base_grads
 
+
+    def predict_proba(self, X):
         kernel = self._compute_kernel(X)
-
-        W_grad = kernel.T @ tau_hat_grad
-        b_grad = tau_hat_grad.sum(0, keepdims=True)
-
-        # Then, add the regularisation grads
-        complete_kernel = self._compute_kernel(self.input_data_)
-        W_grad -= 2 * self.reg * np.dot(complete_kernel, self.W_)
-
-        # Negative sign to force the optimiser to maximise instead of minimise
-        gradients = [-W_grad, -b_grad]
-
-        return gradients
+        return self._infer(kernel)