+++ gemclus/tree/_utils.pyx
@@ -1,5 +1,6 @@
 import numpy as np
 cimport numpy as np
+np.import_array()
 
 cdef class Split: